Fixed but that prevented code from running
Wait command in the thread that displays motor temps to the controller screen wasn't being called. This lead to a thread hogging the cpu, making it impossible for anything else to run. This issue has been solved with an extra wait command that runs even if the main code in the thread isn't running.